Fixed test_plot_data_points to match given data
Some of the expected values in test_plot_data_points have extra digits.  The second data point had a y value of 0.22047244100000002 instead of 0.220472441 as given in the csv data file.  These cause errors with assertEqual.  It works fine with just this change, but it is still checking floats for equality which could be brittle. I also modified the assert line to use np.testing.assert_almost_equal which checks lists in an element-by-element fashion with tolerance and should be more robust.