fixed test_plot_lines
Question asks for regression line "Through the year 2050" but the expected lists stops at the y-value for 2049.  Added datapoint for 2050 in both lines.  assertEqual was throwing errors for differences in 10th digit, changing this to np.testing.assert_almost_equal checks the first 7 digits which is a more robust way to check a regression calculated value.